requirements.txt
matplotlib
numpy
mysql-connector-python
